+++ src/Game/Shibboleth_LuaLoader.cpp
@@ -40,7 +40,7 @@ LuaLoader::~LuaLoader(void)
 {
 }
 
-Gaff::IVirtualDestructor* LuaLoader::load(const char* file_name, unsigned long long, HashMap<AString, IFile*>& file_map)
+Gaff::IVirtualDestructor* LuaLoader::load(const char* file_name, uint64_t, HashMap<AString, IFile*>& file_map)
 {
 	assert(file_map.hasElementWithKey(AString(file_name)));
 	IFile*& file = file_map[AString(file_name)];
