+++ src/Game/include/Shibboleth_ShaderLoader.h
@@ -35,7 +35,7 @@ public:
 	ShaderLoader(RenderManager& render_mgr);
 	~ShaderLoader(void);
 
-	Gaff::IVirtualDestructor* load(const char* file_name, unsigned long long user_data, HashMap<AString, IFile*>& file_map);
+	Gaff::IVirtualDestructor* load(const char* file_name, uint64_t user_data, HashMap<AString, IFile*>& file_map);
 
 private:
 	RenderManager& _render_mgr;
