+++ src/Game/include/Shibboleth_TextureLoader.h
@@ -41,7 +41,7 @@ public:
 	TextureLoader(RenderManager& render_mgr);
 	~TextureLoader(void);
 
-	Gaff::IVirtualDestructor* load(const char* file_name, unsigned long long, HashMap<AString, IFile*>& file_map);
+	Gaff::IVirtualDestructor* load(const char* file_name, uint64_t, HashMap<AString, IFile*>& file_map);
 
 private:
 	RenderManager& _render_mgr;
